fix “get wrong connection project” problem
I found a “get wrong connection project” problem,causes the injection function to be unavailable.

see this log:
“2018-06-19 18:27:23.907189+0800 InjectionIII[61600:2282765] Connection with project file: /Users/zhaolei/Documents/demo/InjectionIII/InjectionIII.xcodeproj”

The reason for the problem : xcode.activeWorkspaceDocument is related to the project opening order，when I open a testProject before  open InjectionIIIProject.And run both xcode(try to connect),xcode.activeWorkspaceDocument point to InjectionIIIProject(can see the above log)

I fixed this problem and have tested it: get connecting project Name from frameworkPath (client write to server), and traverse all workspaceDocuments to find out which project is connecting.

I has 6 years of iOS development experience,but this is my first Github Pull Request.I will be glad that this request can be passed.
And this project is really owesome!