Set indent and tab width on Xcode project
These changes will make it easier for people to contribute while keeping a consistent code style.
Setting the indent and tab width on the Xcode project will override the preference of the developer,
what this means is that developers don't have to change their settings when working on the project.
The settings are set to 4 on both tab and indent to fit with the existing codebase.