Implement TDD feature
Adds a new menu item to enable TDD.

When TDD is enabled, the file watcher will search the project root directory for a corresponding test case for the file that was injected. If a match is found, it will run the tests for that file after it has been injected.

Files are matched using file name using ```containsString:```

The feature is an opt-in feature which means that it is disabled by default. It is enabled or disabled using the applications menu icon.