URL encode projectRoot
This fixes an exception being thrown if the project has spaces in it and the TDD feature is enabled.
Project names need to be properly url encoded before InjectionIII tries to use it. Now the projectRoot
is url encoded using a static method on InjectionServer, it fixes the issue with an exception being
thrown if a URL cannot be created using the current project root.