Improve/print statements (#82)
* Annotate print statements with 💉

Print statements now use 💉 as a prefix. This is done to help the user understand where the message comes from.

* Move Loading .dylib below dlopen

The instruction that the duplicate class warnings should be ignored often get overlooked by the users. The print statement has now been moved to help discoverability of the message. It also uses the same 💉 as the rest of the print messages.

* Add 💉 to "Injection loaded but could not connect"

* Update SwiftEval.swift