Fix crash when sweeping values
First cast the value to `AnyObject` then check that the value isn't `nil`.

This one is a bit iffy because we will get a warning stating that this will always succeed.
In our case, we get `Any` from `Mirror` which doesn't give us any clue if the value is present or not. To work around the non-optional, we try and first cast the object and then check the result as trying to cast the underlying object will end up in failing even if the compiler message tells us otherwise.